third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/text-size-adjust-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'text-size-adjust' to the 'none' keyword: none
  assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
[FAIL] Can set 'text-size-adjust' to a percent: -3.14%
  assert_equals: expected "CSSUnitValue" but got "CSSKeywordValue"
Harness: the test ran to completion.
